requirements.txt
bs4==0.0.1
selenium==2.53.1
PyYAML==3.11
requests==2.9.1
